telstra-id-analyser/python/analyser_handler/requirements.txt
google-cloud-speech==2.8.0
google-cloud-storage==1.42.1
google-cloud-texttospeech==2.5.2